MRG: Only remove original wheel if new wheel is created

We were getting errors when pip downloads a binary wheel, we attempt repair giving a wheel with the same filename, and then delete the original (and therefore, new) wheel on the basis that we have created a new one.

Back off to logic that, if repair generates a wheel with the same filename, we copy the original and discard the repaired version.

+++ manylinux_utils.sh
@@ -10,15 +10,29 @@ function get_platform {
 }
 
 function repair_wheelhouse {
+	# Runs 'auditwheel repair' over all wheels in a directory
+	# If the wheel is not renamed by the repair process,
+	# then the original wheel will be left unmodified
     local in_dir=$1
     local out_dir=${2:-$in_dir}
     for whl in $in_dir/*.whl; do
         if [[ $whl == *none-any.whl ]]; then  # Pure Python wheel
             if [ "$in_dir" != "$out_dir" ]; then cp $whl $out_dir; fi
         else
-            auditwheel repair $whl -w $out_dir/
-            # Remove unfixed if writing into same directory
-            if [ "$in_dir" == "$out_dir" ]; then rm $whl; fi
+            local tmpdir=$(mktemp -d -t)
+            
+            auditwheel repair $whl -w $tmpdir/
+            
+            local built=$(find $tmpdir -name *.whl)
+            if [ $(basename $built) == $(basename $whl) ]; then
+                if [ "$in_dir" != "$out_dir" ]; then cp $whl $out_dir; fi
+            else
+                cp $built $out_dir
+                
+                # Remove unfixed if writing into same directory
+                if [ "$in_dir" == "$out_dir" ]; then rm $whl; fi
+            fi
+            rm -rf $tmpdir
         fi
     done
     chmod -R a+rwX $out_dir
